+++ lib/semantic_release/updaters/gemfile_lock.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module SemanticRelease
+  module Updaters
+    class GemfileLock < BaseUpdater
+      def self.update
+        return if Dir.glob("*.gemspec").empty?
+        return if system("git check-ignore -q Gemfile.lock")
+
+        `bundle check && git add Gemfile.lock`
+      end
+    end
+  end
+end